fixed slow key input response.
It turns out nothing was wrong with the interpreter key handling. It was the fact that the speed of the interpreter
was too slow and it felt jerky, so by increasing the cycle count from 60 Hz to 480 Hz it started to feel like
a fast game. I also changed the mini_fb call to be  rather than  because the latter
would I believe only register keys as they were pressed, while the former was what I was expecting where it would be
return the currently pressed system keys. I also removed the busy loop in  and replaced it with a sleep, which
reduced CPU usage from 100% to ~20%