create Chipotle8 Key enum and use it for keyboard.rs\'s key_input parameter.
Over the next several commits we will get rid of the reliance on minifb::Key and move to our own Key. This is in line
with the idea that chipotle8 is a library, and if you want to use it with a windowing library (such as minifb) then you
will need to implement the appropriate traits to do so. Abstract all the things!