fix timer bugs:
For fun I removed to 2ms sleep in the pong example game loop, thinking it would cause the game to speed ahead. But nothing changed, and it worked just the same.
I'm not sure exactly why, because since the minority of instructions are display ops, it should have caused the game to speed ahead on the non-display ops. Display ops still takes 1-2ms to execute so they would not have slowed down that much. Either way, we don't need this code complicating things, so we remove it. Better yet, it lets use take the CYCLES_PER_SECOND back down to 60 Hz like it should be.